move getTransform function to Lib and use it in bar and treemap
